Simplify service generators
